+++ sticky_mitten_avatar/sma_controller.py
@@ -956,21 +956,20 @@ class StickyMittenAvatarController(Controller):
                            "axis": joint.axis,
                            "avatar_id": self._avatar.id}])
 
-    def rotate_camera_by(self, pitch: float = 0, yaw: float = 0, roll: float = 0) -> None:
+    def rotate_camera_by(self, pitch: float = 0, yaw: float = 0) -> None:
         """
         Rotate an avatar's camera around each axis.
-        The head of the avatar won't visually rotate (as this could put the avatar off-balance).
+        The head of the avatar won't visually rotate, as this could put the avatar off-balance.
         Advances the simulation by 1 frame.
 
         :param pitch: Pitch (nod your head "yes") the camera by this angle, in degrees.
         :param yaw: Yaw (shake your head "no") the camera by this angle, in degrees.
-        :param roll: Roll (put your ear to your shoulder) the camera by this angle, in degrees.
         """
 
         self._start_task()
 
         commands = []
-        for angle, axis in zip([pitch, yaw, roll], ["pitch", "yaw", "roll"]):
+        for angle, axis in zip([pitch, yaw], ["pitch", "yaw"]):
             commands.append({"$type": "rotate_sensor_container_by",
                              "axis": axis,
                              "angle": angle,