+++ sticky_mitten_avatar/sma_controller.py
@@ -181,10 +181,14 @@ class StickyMittenAvatarController(FloorplanController):
         from sticky_mitten_avatar import StickyMittenAvatarController
 
         c = StickyMittenAvatarController()
-        c.init_scene(scene="floorplan_3", layout=0)
+        c.init_scene(scene="floorplan_2b", layout=0)
         ```
 
-        For valid scenes and layouts, [read this](https://github.com/threedworld-mit/tdw/blob/master/Documentation/python/floorplan_controller.md).
+        Valid scenes and layouts:
+
+        | `scene` | `layout` |
+        | --- | --- |
+        | `"2a"`, `"2b"`, or `"2b"` | 0 |
 
         :param scene: The name of an interior floorplan scene. If None, the controller will load a simple empty room.
         :param layout: The furniture layout of the floorplan. If None, the controller will load a simple empty room.