+++ changelog.md
@@ -22,6 +22,7 @@
   - Fixed: No audio data for sub-objects of composite objects.
 - Added utility script: `composite_object_audio.py` Get default audio parameters for sub-objects of composite objects.
 - Added test controller: `composite_object_test.py` Test if the avatar can grasp a sub-object of a composite object.
+- Added: `composite_object_audio.json` Audio values for every sub-object of a composite object.
 
 ## 0.3.2