+++ changelog.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 0.4.0
+
+### Frontend
+
+- `StickyMittenAvatarController`:
+  - **Added optional parameters `scene` and `layout` to `init_scene()` to load a floorplan and layout.** 
+- `FrameData`:
+  - Reorganized the fields in the documentation and added more and better code examples.
+
+### Backend
+
+- `StickyMittenAvatarController`:
+  - Now a subclass of `FloorplanController` (was a subclass of `Controller`).
+  - Objects are initialized via `AudioInitData` and `ObjectInitData`.
+  - Removed friction dictionaries (handled in `AudioInitData`).
+  - Renamed `_get_scene_init_commands_early()` to `_get_scene_init_commands()`.
+  - Removed: `_do_scene_init_late()`.
+- `FrameData`:
+  - Fixed: Unhandled exception if there is a collision between objects not yet in the static object info dictionary (this is sometimes possible with composite objects).
+- Revised `__init__.py` to make it harder to accidentally use `TestController` and `BoxRoomContainers`.
+- Removed a lot of audio values (they are now in the `tdw` repo).
+- Added composite object audio data for `jigsaw_puzzle_composite` and `rope_table_lamp`.
+- Fixed: `composite_object_audio.py` doesn't use the correct sub-object ID.
+- Removed: `init_commands.py` (not needed).
+
 ## 0.3.3
 
 ### Frontend